cargo fmt, split files and refactor...
